examples/semantic-kernel/requirements.txt
semantic-kernel>=0.9.0
sentinel-client>=0.1.0
python-dotenv
